Create NaiveBayes.ipynb
3.3.1 Choosing the Best Classifier Configuration
In this part of the assignment, the objective is to select the best possible classifier configuration. First, divide the training data into two subsets: training data and validation data, in an 80-20 or 90-10 ratio. The first subset will be used to train the different classifier variants, and the second subset will be used to evaluate and compare the results.

In this practice, you are asked to train the following configurations:

Test the performance of the Naive Bayes classifier with Bernoulli distribution and Multinomial distribution.
Compare the results of both classifiers for different values of the Laplace smoothing hyperparameter. Note that it is possible to use values greater than or less than 1.
Analyze how the use of bi-grams as features in the bag-of-words influences the classifier.
To determine the best classifier, the F1-score will be used, obtained through predictions on the validation data.

3.3.2 Evaluation of the Best Classifier on Test Data
Once you have chosen the classifier and the hyperparameters that work best, the final part of the assignment will focus on a more detailed analysis of the results using the test data and the various metrics discussed in class: F1-Score, confusion matrix, and precision-recall curve.

Interpret the results obtained for your classifier. Do you consider it to be a good spam email classifier?
Select an appropriate threshold for the spam classifier based on the precision-recall curve, justifying your response.
Select some examples of spam and ham emails from the test set that were classified correctly and incorrectly, and discuss the results.