Pattern recognition with Neural networks
The general objective is to solve real classification problems using neural networks with Keras, a simple API for Neural Networks that allows training and executing neural networks using TensorFlow as a back-end. Just like in practice 4, Python will be used within the Google Colab environment.

The goal of the exercise is to find a network that has the lowest possible classification error, using accuracy as the comparison metric. Start building a table that reflects the different networks you have tested and the results they achieve. As part of the documentation, you should present this table, discussed and commented on appropriately.

Although the goal of the work is to achieve the highest possible accuracy, what will be most considered in the evaluation is the design process and the interpretation of the different results obtained, rather than reaching a specific accuracy value.