Image classifier with convolutional neural network
The general objective is to solve real classification problems using convolutional neural networks with Keras, similar to neural networks from before (Python will also be used within the Google Colab environment).

The goal of this practice is to train a convolutional neural network and compare its results with those obtained by an MLP network trained for the same task. (Make sure it has the lowest possible classification error, using accuracy as the comparison metric. Just like in Practice 6, build a table reflecting the variations you have tried and the results they obtain. As part of the documentation, you must present this table, discussed and adequately commented on.)

Although the aim of the work is to achieve the highest possible accuracy, what will be most taken into account for evaluation is the design process and the interpretation of the different results obtained, not the fact of reaching a specific accuracy value.